refactor: consolidate core into binary package
