+++ Cargo.toml
@@ -1,7 +1,30 @@
+[package]
+name = "rpc-plane"
+description = "Solana RPC proxy — intelligent multi-provider routing"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+readme = "README.md"
+keywords = ["solana", "rpc", "proxy", "failover"]
+categories = ["command-line-utilities", "network-programming"]
+autolib = false
+include = [
+    "/src/**",
+    "/benches/**",
+    "/build.rs",
+    "/Cargo.toml",
+    "/README.md",
+    "/LICENSE",
+    "/config.example.toml",
+]
+
+[[bin]]
+name = "rpc-plane"
+path = "src/main.rs"
+
 [workspace]
 members = [
-    "rpc-plane",
-    "rpc-plane-core",
     "tools/dummy-rpc",
     "tools/load-test",
 ]
@@ -42,3 +65,38 @@ hyper = { version = "1", features = ["http1", "server"] }
 hyper-util = { version = "0.1", features = ["tokio", "server-auto", "server-graceful"] }
 tower = "0.5"
 http-body-util = "0.1"
+
+[dependencies]
+tokio.workspace = true
+axum.workspace = true
+reqwest.workspace = true
+bytes.workspace = true
+base64.workspace = true
+bs58.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+toml.workspace = true
+tracing.workspace = true
+tracing-subscriber.workspace = true
+clap.workspace = true
+anyhow.workspace = true
+uuid.workspace = true
+prometheus.workspace = true
+parking_lot.workspace = true
+rand.workspace = true
+shellexpand.workspace = true
+tower-http.workspace = true
+libc.workspace = true
+hyper.workspace = true
+hyper-util.workspace = true
+tower.workspace = true
+
+[dev-dependencies]
+tempfile.workspace = true
+tower = { workspace = true, features = ["util"] }
+http-body-util.workspace = true
+criterion = { version = "0.5", features = ["html_reports"] }
+
+[[bench]]
+name = "proxy_bench"
+harness = false
