--- rpc-plane-core/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "rpc-plane-core"
-description = "Core proxy logic for rpc-plane"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-tokio.workspace = true
-axum.workspace = true
-reqwest.workspace = true
-bytes.workspace = true
-base64.workspace = true
-bs58.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-toml.workspace = true
-tracing.workspace = true
-anyhow.workspace = true
-thiserror.workspace = true
-uuid.workspace = true
-prometheus.workspace = true
-parking_lot.workspace = true
-rand.workspace = true
-shellexpand.workspace = true
-tower-http.workspace = true
-
-[dev-dependencies]
-tempfile.workspace = true
-tower = { version = "0.5", features = ["util"] }
-tower-http = { workspace = true }
-http-body-util = "0.1"
-criterion = { version = "0.5", features = ["html_reports"] }
-
-[[bench]]
-name = "proxy_bench"
-harness = false
