--- rpc-plane-core/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod config;
-pub mod health;
-pub mod metrics;
-pub mod proxy;
-pub mod router;
-pub mod telemetry;
-pub mod tx;
