--- rpc-plane/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "rpc-plane"
-description = "Solana RPC proxy — intelligent multi-provider routing"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[[bin]]
-name = "rpc-plane"
-path = "src/main.rs"
-
-
-[dependencies]
-rpc-plane-core = { path = "../rpc-plane-core" }
-tokio.workspace = true
-axum.workspace = true
-reqwest.workspace = true
-serde_json.workspace = true
-tracing.workspace = true
-tracing-subscriber.workspace = true
-clap.workspace = true
-anyhow.workspace = true
-bytes.workspace = true
-parking_lot.workspace = true
-libc.workspace = true
-serde.workspace = true
-toml.workspace = true
-hyper.workspace = true
-hyper-util.workspace = true
-tower.workspace = true
